Replace 2 functions to utils.py
