+++ personal_helper/commands.py
@@ -351,35 +351,3 @@ def print_contacts(addressbook: AB) -> None:
     table_ful = TablePrinter(table)
     table_ful.print_table()
 
-
-def get_validation_commands(commands: list, target_command: str) -> dict:
-    result = {}
-    
-    for command in commands:
-        result[command] = []
-        for index, char in enumerate(command):
-            try:
-                result[command].append(target_command[index] == char)
-            except IndexError:
-                pass
-        
-    return result
-
-def normalize_command(D: dict) -> str:
-    
-    max_true_count = 0
-    max_true_element = None
-
-    for element, values in D.items():
-        true_count = values.count(True)
-
-        if true_count > max_true_count:
-            max_true_count = true_count
-            max_true_element = element
-
-    
-
-    if max_true_element is None:
-        return None
-    
-    return f"{max_true_element}"
